mod: code review
